Move games into own components folder except for simon
