+++ client/src/components/Games/Scramble/Score.js
