+++ client/src/components/Games/Scramble/Timer.js
